requirements.txt
mss==9.0.1
Pillow==10.3.0
PyQt6==6.7.0
keyboard==0.13.5